+++ src/Silex/Provider/FormServiceProvider.php
@@ -20,6 +20,9 @@ use Symfony\Component\Form\Extension\HttpFoundation\HttpFoundationExtension;
 use Symfony\Component\Form\Extension\Validator\ValidatorExtension as FormValidatorExtension;
 use Symfony\Component\Form\Forms;
 use Symfony\Component\Form\ResolvedFormTypeFactory;
+use Symfony\Component\Security\Csrf\CsrfTokenManager;
+use Symfony\Component\Security\Csrf\TokenStorage\NativeSessionTokenStorage;
+use Symfony\Component\Security\Csrf\TokenStorage\SessionTokenStorage;
 
 /**
  * Symfony Form component Provider.
@@ -78,7 +81,10 @@ class FormServiceProvider implements ServiceProviderInterface
 
                 if (isset($app['translator']) && method_exists($app['translator'], 'addResource')) {
                     $r = new \ReflectionClass('Symfony\Component\Form\Form');
-                    $app['translator']->addResource('xliff', dirname($r->getFilename()).'/Resources/translations/validators.'.$app['locale'].'.xlf', $app['locale'], 'validators');
+                    $file = dirname($r->getFilename()).'/Resources/translations/validators.'.$app['locale'].'.xlf';
+                    if (file_exists($file)) {
+                        $app['translator']->addResource('xliff', $file, $app['locale'], 'validators');
+                    }
                 }
             }
 
@@ -101,11 +107,19 @@ class FormServiceProvider implements ServiceProviderInterface
         };
 
         $app['form.csrf_provider'] = function ($app) {
-            if (isset($app['session'])) {
-                return new SessionCsrfProvider($app['session'], $app['form.secret']);
-            }
+            if (!class_exists('Symfony\Component\Form\Extension\DataCollector\DataCollectorExtension')) {
+                // Symfony 2.3
+                if (isset($app['session'])) {
+                    return new SessionCsrfProvider($app['session'], $app['form.secret']);
+                }
 
-            return new DefaultCsrfProvider($app['form.secret']);
+                return new DefaultCsrfProvider($app['form.secret']);
+            } else {
+                // Symfony 2.4+
+                $storage = isset($app['session']) ? new SessionTokenStorage($app['session']) : new NativeSessionTokenStorage();
+
+                return new CsrfTokenManager(null, $storage);
+            }
         };
     }
 }