+++ src/Silex/Provider/HttpCacheServiceProvider.php
@@ -19,6 +19,7 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpKernel\HttpCache\Esi;
 use Symfony\Component\HttpKernel\HttpCache\Store;
 use Symfony\Component\HttpKernel\EventListener\EsiListener;
+use Symfony\Component\HttpKernel\EventListener\SurrogateListener;
 
 /**
  * Symfony HttpKernel component Provider for HTTP cache.
@@ -48,6 +49,10 @@ class HttpCacheServiceProvider implements ServiceProviderInterface, EventListene
         };
 
         $app['http_cache.esi_listener'] = function ($app) {
+            if (class_exists('Symfony\Component\HttpKernel\EventListener\SurrogateListener')) {
+                return new SurrogateListener($app['http_cache.esi']);
+            }
+
             return new EsiListener($app['http_cache.esi']);
         };