+++ src/Silex/Provider/ValidatorServiceProvider.php
@@ -14,10 +14,13 @@ namespace Silex\Provider;
 use Pimple\Container;
 use Pimple\ServiceProviderInterface;
 use Silex\Provider\Validator\ConstraintValidatorFactory;
+use Silex\Application;
 use Symfony\Component\Validator\Validator;
 use Symfony\Component\Validator\DefaultTranslator;
 use Symfony\Component\Validator\Mapping\ClassMetadataFactory;
+use Symfony\Component\Validator\Mapping\Factory\LazyLoadingMetadataFactory;
 use Symfony\Component\Validator\Mapping\Loader\StaticMethodLoader;
+use Symfony\Component\Validator\Validation;
 
 /**
  * Symfony Validator component Provider.
@@ -29,22 +32,35 @@ class ValidatorServiceProvider implements ServiceProviderInterface
     public function register(Container $app)
     {
         $app['validator'] = function ($app) {
-            if (isset($app['translator']) && method_exists($app['translator'], 'addResource')) {
-                $r = new \ReflectionClass('Symfony\Component\Validator\Validator');
-
-                $app['translator']->addResource('xliff', dirname($r->getFilename()).'/Resources/translations/validators.'.$app['locale'].'.xlf', $app['locale'], 'validators');
+            if (isset($app['translator'])) {
+                $r = new \ReflectionClass('Symfony\Component\Validator\Validation');
+                $file = dirname($r->getFilename()).'/Resources/translations/validators.'.$app['locale'].'.xlf';
+                if (file_exists($file)) {
+                    $app['translator']->addResource('xliff', $file, $app['locale'], 'validators');
+                }
             }
 
-            return new Validator(
-                $app['validator.mapping.class_metadata_factory'],
-                $app['validator.validator_factory'],
-                isset($app['translator']) ? $app['translator'] : new DefaultTranslator(),
-                'validators',
-                $app['validator.object_initializers']
-            );
+            return $app['validator.builder']->getValidator();
+        };
+
+        $app['validator.builder'] = function ($app) {
+            $builder = Validation::createValidatorBuilder();
+            $builder->setConstraintValidatorFactory($app['validator.validator_factory']);
+            $builder->setTranslationDomain('validators');
+            $builder->addObjectInitializers($app['validator.object_initializers']);
+            $builder->setMetadataFactory($app['validator.mapping.class_metadata_factory']);
+            if (isset($app['translator'])) {
+                $builder->setTranslator($app['translator']);
+            }
+
+            return $builder;
         };
 
         $app['validator.mapping.class_metadata_factory'] = function ($app) {
+            if (class_exists('Symfony\Component\Validator\Mapping\Factory\LazyLoadingMetadataFactory')) {
+                return new LazyLoadingMetadataFactory(new StaticMethodLoader());
+            }
+
             return new ClassMetadataFactory(new StaticMethodLoader());
         };