+++ tests/Silex/Tests/Provider/FormServiceProviderTest.php
@@ -14,12 +14,17 @@ namespace Silex\Tests\Provider;
 use Silex\Application;
 use Silex\Provider\FormServiceProvider;
 use Silex\Provider\TranslationServiceProvider;
+use Silex\Provider\ValidatorServiceProvider;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\Extension\Csrf\CsrfProvider\CsrfProviderInterface;
 use Symfony\Component\Form\FormTypeGuesserChain;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
+use Symfony\Component\Security\Csrf\CsrfToken;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 class FormServiceProviderTest extends \PHPUnit_Framework_TestCase
 {
@@ -110,7 +115,35 @@ class FormServiceProviderTest extends \PHPUnit_Framework_TestCase
         ))));
 
         $this->assertFalse($form->isValid());
-        $this->assertContains('ERROR: German translation', $form->getErrorsAsString());
+        $r = new \ReflectionMethod($form, 'getErrors');
+        if (!$r->getNumberOfParameters()) {
+            $this->assertContains('ERROR: German translation', $form->getErrorsAsString());
+        } else {
+            // as of 2.5
+            $this->assertContains('ERROR: German translation', (string) $form->getErrors(true, false));
+        }
+    }
+
+    public function testFormServiceProviderWillNotAddNonexistentTranslationFiles()
+    {
+        $app = new Application(array(
+            'locale' => 'nonexistent',
+        ));
+
+        $app->register(new FormServiceProvider());
+        $app->register(new ValidatorServiceProvider());
+        $app->register(new TranslationServiceProvider(), array(
+            'locale_fallbacks' => array(),
+        ));
+
+        $app['form.factory'];
+        $translator = $app['translator'];
+
+        try {
+            $translator->trans('test');
+        } catch (NotFoundResourceException $e) {
+            $this->fail('Form factory should not add a translation resource that does not exist');
+        }
     }
 }
 
@@ -125,28 +158,72 @@ class DummyFormType extends AbstractType
     }
 }
 
-class DummyFormTypeExtension extends AbstractTypeExtension
-{
-    public function getExtendedType()
+if (method_exists('Symfony\Component\Form\AbstractType', 'configureOptions')) {
+    class DummyFormTypeExtension extends AbstractTypeExtension
     {
-        return 'file';
-    }
+        public function getExtendedType()
+        {
+            return 'file';
+        }
 
-    public function setDefaultOptions(OptionsResolverInterface $resolver)
+        public function configureOptions(OptionsResolver $resolver)
+        {
+            $resolver->setDefined(array('image_path'));
+        }
+    }
+} else {
+    class DummyFormTypeExtension extends AbstractTypeExtension
     {
-        $resolver->setOptional(array('image_path'));
+        public function getExtendedType()
+        {
+            return 'file';
+        }
+
+        public function setDefaultOptions(OptionsResolverInterface $resolver)
+        {
+            if (!method_exists($resolver, 'setDefined')) {
+                $resolver->setOptional(array('image_path'));
+            } else {
+                $resolver->setDefined(array('image_path'));
+            }
+        }
     }
 }
 
-class FakeCsrfProvider implements CsrfProviderInterface
-{
-    public function generateCsrfToken($intention)
+if (!class_exists('Symfony\Component\Form\Extension\DataCollector\DataCollectorExtension')) {
+    // Symfony 2.3 only
+    class FakeCsrfProvider implements CsrfProviderInterface
     {
-        return $intention.'123';
-    }
+        public function generateCsrfToken($intention)
+        {
+            return $intention.'123';
+        }
 
-    public function isCsrfTokenValid($intention, $token)
+        public function isCsrfTokenValid($intention, $token)
+        {
+            return $token === $this->generateCsrfToken($intention);
+        }
+    }
+} else {
+    class FakeCsrfProvider implements CsrfTokenManagerInterface
     {
-        return $token === $this->generateCsrfToken($intention);
+        public function getToken($tokenId)
+        {
+            return new CsrfToken($tokenId, '123');
+        }
+
+        public function refreshToken($tokenId)
+        {
+            return new CsrfToken($tokenId, '123');
+        }
+
+        public function removeToken($tokenId)
+        {
+        }
+
+        public function isTokenValid(CsrfToken $token)
+        {
+            return '123' === $token->getValue();
+        }
     }
 }