+++ tests/Silex/Tests/Provider/MonologServiceProviderTest.php
@@ -41,7 +41,9 @@ class MonologServiceProviderTest extends \PHPUnit_Framework_TestCase
 
         $this->assertTrue($app['monolog.handler']->hasDebug('> GET /foo'));
         $this->assertTrue($app['monolog.handler']->hasDebug('< 200'));
-        $this->assertTrue($app['monolog.handler']->hasInfo('Matched route "GET_foo" (parameters: "_controller": "{}", "_route": "GET_foo")'));
+
+        $records = $app['monolog.handler']->getRecords();
+        $this->assertContains('Matched route "GET_foo"', $records[0]['message']);
     }
 
     public function testManualLogging()