+++ tests/Silex/Tests/Provider/SessionServiceProviderTest.php
@@ -94,7 +94,7 @@ class SessionServiceProviderTest extends WebTestCase
         });
 
         $app['debug'] = true;
-        $app['exception_handler']->disable();
+        unset($app['exception_handler']);
 
         $client = new Client($app);