+++ tests/Silex/Tests/Provider/ValidatorServiceProviderTest.php
@@ -12,8 +12,10 @@
 namespace Silex\Tests\Provider;
 
 use Silex\Application;
+use Silex\Provider\TranslationServiceProvider;
 use Silex\Provider\ValidatorServiceProvider;
 use Silex\Provider\FormServiceProvider;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 use Symfony\Component\Validator\Constraints as Assert;
 use Silex\Tests\Provider\ValidatorServiceProviderTest\Constraint\Custom;
 use Silex\Tests\Provider\ValidatorServiceProviderTest\Constraint\CustomValidator;
@@ -67,7 +69,7 @@ class ValidatorServiceProviderTest extends \PHPUnit_Framework_TestCase
      */
     public function testValidatorServiceIsAValidator($app)
     {
-        $this->assertInstanceOf('Symfony\Component\Validator\Validator', $app['validator']);
+        $this->assertInstanceOf('Symfony\Component\Validator\ValidatorInterface', $app['validator']);
     }
 
     /**
@@ -93,11 +95,32 @@ class ValidatorServiceProviderTest extends \PHPUnit_Framework_TestCase
             ->getForm()
         ;
 
-        $form->bind(array('email' => $email));
+        $form->submit(array('email' => $email));
 
         $this->assertEquals($isValid, $form->isValid());
         $this->assertEquals($nbGlobalError, count($form->getErrors()));
         $this->assertEquals($nbEmailError, count($form->offsetGet('email')->getErrors()));
+    }
+
+    public function testValidatorWillNotAddNonexistentTranslationFiles()
+    {
+        $app = new Application(array(
+            'locale' => 'nonexistent',
+        ));
+
+        $app->register(new ValidatorServiceProvider());
+        $app->register(new TranslationServiceProvider(), array(
+            'locale_fallbacks' => array(),
+        ));
+
+        $app['validator'];
+        $translator = $app['translator'];
+
+        try {
+            $translator->trans('test');
+        } catch (NotFoundResourceException $e) {
+            $this->fail('Validator should not add a translation resource that does not exist');
+        }
     }
 
     public function testValidatorConstraintProvider()